% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_apikey.R
\name{assign_apikey}
\alias{assign_apikey}
\title{Allows you to use the inputted api key (from alphavantage) for this package}
\usage{
assign_apikey(apikey)
}
\arguments{
\item{apikey}{Your apikey from alphavantage}
}
\value{
NULL, but the apikey is saved for the usage of the package
}
\description{
Allows you to use the inputted api key (from alphavantage) for this package
}
