% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Given an interval of dates and a company's stock symbol, this function will give you the
opening, highest, lowest, closing, or volume stock data.
The function allows you to view the past 20 years of stock data.}
\usage{
get_data(from, to, stock_symbol, view_type, api_key = apikey)
}
\arguments{
\item{from}{The starting date you would like to view}

\item{to}{The ending date you would like to view}

\item{stock_symbol}{The stock symbol of a company}

\item{view_type}{An integer that represents 1 = open, 2 = high, 3 = low, 4 = close, or 5 = volume stock data}

\item{apikey}{Default is to use the apikey using the assign_apikey function}
}
\value{
A dataframe of one column with the dates and another column with the stock data
}
\description{
Given an interval of dates and a company's stock symbol, this function will give you the
opening, highest, lowest, closing, or volume stock data.
The function allows you to view the past 20 years of stock data.
}
